# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cell Messenger API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Cell API Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Прямое подключение к Go серверу (локально или контейнер API)"
    },
    {
      "url": "http://localhost",
      "description": "Подключение через Nginx Reverse Proxy"
    },
    {
      "url": "/",
      "description": "Относительный корень сервера"
    }
  ],
...
  "paths": {
...
    "/api/v1/calls/rooms": {
      "post": {
        "tags": [
          "Calls"
        ],
        "summary": "Создание или получение активной комнаты звонка для чата",
        "description": "Создает новую комнату для аудио/видео звонка в указанном чате или возвращает уже существующую активную комнату. Доступно только участникам чата.",
        "operationId": "createCallRoom",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCallRoomRequest"
              },
              "example": {
                "chat_id": "550e8400-e29b-41d4-a716-446655440000",
                "type": "video"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Комната успешно создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallRoomResponse"
                }
              }
            }
          },
          "200": {
            "description": "Комната уже существует и активна в данном чате",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallRoomResponse"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос или отсутствие chat_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Пользователь не является участником данного чата",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Авторизационный JWT токен. Передается в заголовке `Authorization: Bearer <token>`"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Описание ошибки",
            "example": "invalid request payload"
          },
          "sys": {
            "type": "string",
            "description": "Системное/детализированное сообщение об ошибке (при наличии)",
            "example": "context deadline exceeded"
          }
        }
      },
...
      "CreateCallRoomRequest": {
        "type": "object",
        "required": [
          "chat_id"
        ],
        "properties": {
          "chat_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID чата, в котором инициируется звонок",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "type": {
            "type": "string",
            "enum": [
              "video",
              "audio"
            ],
            "default": "video",
            "description": "Тип звонка: видео или аудио",
            "example": "video"
          }
        }
      },
      "CallParticipant": {
        "type": "object",
        "required": [
          "peer_id",
          "user_id",
          "username",
          "joined_at"
        ],
        "properties": {
          "peer_id": {
            "type": "string",
            "description": "Уникальный идентификатор пира в сессии SFU",
            "example": "a8f3b2c1d4e5"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID пользователя",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "username": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "alex_dev"
          },
          "joined_at": {
            "type": "string",
            "format": "date-time",
            "description": "Время подключения к комнате",
            "example": "2026-10-05T07:30:00Z"
          }
        }
      },
      "CallRoomResponse": {
        "type": "object",
        "required": [
          "room_id",
          "chat_id",
          "call_type",
          "initiator_id",
          "created_at",
          "is_active",
          "participants"
        ],
        "properties": {
          "room_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID комнаты звонка",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "chat_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID чата",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "call_type": {
            "type": "string",
            "enum": [
              "video",
              "audio"
            ],
            "description": "Тип звонка",
            "example": "video"
          },
          "initiator_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID пользователя-инициатора",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "initiator_name": {
            "type": "string",
            "description": "Имя инициатора звонка",
            "example": "alex_dev"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Время создания комнаты",
            "example": "2026-10-05T07:30:00Z"
          },
          "is_active": {
            "type": "boolean",
            "description": "Флаг активности комнаты звонка",
            "example": true
          },
          "participants": {
            "type": "array",
            "description": "Список текущих участников звонка",
            "items": {
              "$ref": "#/components/schemas/CallParticipant"
            }
          }
        }
      },
...
    }
  }
}